task 0: prints all elements of linked list

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -7,11 +7,11 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	unsigned long int count = 0;
+	size_t count = 0;
 
 	while (h)
 	{
-		printf("%lu\n", h->n);
+		printf("%d\n", h->n);
 		count++;
 		h = h->next;
 	}
