+++ 0x13-more_singly_linked_lists/lists.h
@@ -12,7 +12,7 @@
  */
 typedef struct list
 {
-	unsigned long int n;
+	int n;
 	struct list *next;
 } listint_t;
